retrieval_gen/TitleGen/requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
attrs==25.1.0
beautifulsoup4==4.13.3
cairocffi==1.7.1
CairoSVG==2.7.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
cssselect2==0.7.0
defusedxml==0.7.1
distro==1.9.0
docopt==0.6.2
exceptiongroup==1.2.2
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
Js2Py==0.74
multidict==6.1.0
numpy==2.2.3
openai==1.63.0
opencv-python==4.11.0.86
packaging==24.2
pillow==11.1.0
pipwin==0.5.2
propcache==0.2.1
pycairo==1.27.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyjsparser==2.7.1
PyPrind==2.11.3
pySmartDL==1.3.4
requests==2.32.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
svgwrite==1.4.3
tinycss2==1.4.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3
urllib3==2.3.0
webencodings==0.5.1
yarl==1.18.3
